+++ src/main/java/com/github/standobyte/jojo/client/entityrender/stand/StandEntityModel.java
@@ -210,8 +210,7 @@ public class StandEntityModel<T extends StandEntity, S extends StandEntityRender
 		Reminder.thatThisShouldBeInAnEntityModelMixinInstead();
 		if (shouldRenderBarrageLayer(renderState)) {
 			BarrageSwings.currentlyRendering.renderLayerBarrage((EntityModel<?>) (Object) this, 
-					poseStack, com.github.standobyte.jojo.client.rendertype.BarrageVertexConsumer.forBarrage(buffer),
-					packedLight, packedOverlay, renderColor,
+					poseStack, buffer, packedLight, packedOverlay, renderColor,
 					renderState != null ? renderState.xRot : 0);
 		}
 	}